refactor: remove old code implementation
